audi_api/requirements.txt
openpyxl==3.0.7
virtualenv==16.7.8
requests==2.23.0
python-dotenv==0.13.0
Flask==1.1.1
Flask-Cors==3.0.8
